Bugs fixed

+++ controllers/charactersController.js
@@ -26,25 +26,34 @@ exports.getCharacterById = async (req, res) => {
 
     res.json(character);
   } catch (err) {
-    res.status(500).json({ message: 'Failed to fetch character' });
+    res.status(500).json({ message: 'Validation failed', error: err.message });
   }
 };
 
 exports.createCharacter = async (req, res) => {
+
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
   try {
     const character = new Character(req.body);
     await character.save();
     res.status(201).json(character);
   } catch (err) {
-    res.status(400).json({ message: err.message });
+    res.status(500).json({ message: err.message });
   }
 };
 
 // UPDATE (PUT)
 exports.updateCharacter = async (req, res) => {
   try {
+    const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: 'Invalid ID format' });
+    }
     const character = await Character.findByIdAndUpdate(
-      req.params.id,
+      id,
       req.body,
       { new: true, runValidators: true }
     );
@@ -54,13 +63,17 @@ exports.updateCharacter = async (req, res) => {
     }
     res.status(200).json(character);
   } catch (err) {
-    res.status(400).json({ message: err.message });
+    res.status(400).json({ message:'Update failed', error: err.message });
   }
 };
 
 // DELETE
 exports.deleteCharacter = async (req, res) => {
   try {
+    const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: 'Invalid ID format' });
+    }
     const character = await Character.findByIdAndDelete(req.params.id);
 
     if (!character) {
@@ -69,6 +82,6 @@ exports.deleteCharacter = async (req, res) => {
 
     res.status(200).json({ message: 'Character deleted successfully' });
   } catch (err) {
-    res.status(400).json({ message: 'Invalid ID format' });
+    res.status(500).json({ message: 'Delete failed', error: err.message });
   }
 };
